third_party/blink/web_tests/external/wpt/html/semantics/forms/the-select-element/select-value-expected.txt

This is a testharness.js-based test.
[FAIL] option is child of div
  assert_equals: expected "" but got "1"
Harness: the test ran to completion.
